refactor: reorganize Smartsheet API into modular TypeScript classes
